[IceCream] refactor SyncEngine
