+++ IceCream/Classes/DatabaseManager.swift
@@ -1,0 +1,137 @@
+//
+//  DatabaseManager.swift
+//  IceCream
+//
+//  Created by caiyue on 2019/4/22.
+//
+
+import CloudKit
+
+public protocol DatabaseManager: class {
+    
+    /// A conduit for accessing and performing operations on the data of an app container.
+    var database: CKDatabase { get }
+    
+    /// An encapsulation of content associated with an app.
+    var container: CKContainer { get }
+    
+    var syncObjects: [Syncable] { get }
+    
+    init(objects: [Syncable], container: CKContainer)
+    
+    func prepare()
+    
+    func fetchChangesInDatabase(_ callback: (() -> Void)?)
+    
+    /// The CloudKit Best Practice is out of date, now use this:
+    /// https://developer.apple.com/documentation/cloudkit/ckoperation
+    /// Which problem does this func solve? E.g.:
+    /// 1.(Offline) You make a local change, involve a operation
+    /// 2. App exits or ejected by user
+    /// 3. Back to app again
+    /// The operation resumes! All works like a magic!
+    func resumeLongLivedOperationIfPossible()
+    
+    func createCustomZonesIfAllowed(_ completion: ((Error?) -> ())?)
+    func startObservingRemoteChanges()
+    func startObservingTermination()
+    func createDatabaseSubscriptionIfHaveNot()
+    func registerLocalDatabase()
+    
+    func cleanUp()
+}
+
+extension DatabaseManager {
+    
+    func prepare() {
+        syncObjects.forEach {
+            $0.pipeToEngine = { [weak self] recordsToStore, recordIDsToDelete in
+                guard let self = self else { return }
+                self.syncRecordsToCloudKit(recordsToStore: recordsToStore, recordIDsToDelete: recordIDsToDelete)
+            }
+        }
+    }
+    
+    func resumeLongLivedOperationIfPossible() {
+        container.fetchAllLongLivedOperationIDs { [weak self]( opeIDs, error) in
+            guard let self = self else { return }
+            guard error == nil else { return }
+            guard let ids = opeIDs else { return }
+            for id in ids {
+                self.container.fetchLongLivedOperation(withID: id, completionHandler: { [weak self](ope, error) in
+                    guard let self = self else { return }
+                    guard error == nil else { return }
+                    if let modifyOp = ope as? CKModifyRecordsOperation {
+                        modifyOp.modifyRecordsCompletionBlock = { (_,_,_) in
+                            print("Resume modify records success!")
+                        }
+                        self.container.add(modifyOp)
+                    }
+                })
+            }
+        }
+    }
+    
+    func startObservingRemoteChanges() {
+        NotificationCenter.default.addObserver(forName: Notifications.cloudKitDataDidChangeRemotely.name, object: nil, queue: nil, using: { [weak self](_) in
+            guard let self = self else { return }
+            DispatchQueue.global(qos: .utility).async {
+                self.fetchChangesInDatabase(nil)
+            }
+        })
+    }
+    
+    /// Sync local data to CloudKit
+    /// For more about the savePolicy: https://developer.apple.com/documentation/cloudkit/ckrecordsavepolicy
+    public func syncRecordsToCloudKit(recordsToStore: [CKRecord], recordIDsToDelete: [CKRecord.ID], completion: ((Error?) -> ())? = nil) {
+        let modifyOpe = CKModifyRecordsOperation(recordsToSave: recordsToStore, recordIDsToDelete: recordIDsToDelete)
+        
+        if #available(iOS 11.0, OSX 10.13, tvOS 11.0, watchOS 4.0, *) {
+            let config = CKOperation.Configuration()
+            config.isLongLived = true
+            modifyOpe.configuration = config
+        } else {
+            // Fallback on earlier versions
+            modifyOpe.isLongLived = true
+        }
+        
+        // We use .changedKeys savePolicy to do unlocked changes here cause my app is contentious and off-line first
+        // Apple suggests using .ifServerRecordUnchanged save policy
+        // For more, see Advanced CloudKit(https://developer.apple.com/videos/play/wwdc2014/231/)
+        modifyOpe.savePolicy = .changedKeys
+        
+        // To avoid CKError.partialFailure, make the operation atomic (if one record fails to get modified, they all fail)
+        // If you want to handle partial failures, set .isAtomic to false and implement CKOperationResultType .fail(reason: .partialFailure) where appropriate
+        modifyOpe.isAtomic = true
+        
+        modifyOpe.modifyRecordsCompletionBlock = {
+            [weak self]
+            (_, _, error) in
+            
+            guard let self = self else { return }
+            
+            switch ErrorHandler.shared.resultType(with: error) {
+            case .success:
+                DispatchQueue.main.async {
+                    completion?(nil)
+                }
+            case .retry(let timeToWait, _):
+                ErrorHandler.shared.retryOperationIfPossible(retryAfter: timeToWait) {
+                    self.syncRecordsToCloudKit(recordsToStore: recordsToStore, recordIDsToDelete: recordIDsToDelete, completion: completion)
+                }
+            case .chunk:
+                /// CloudKit says maximum number of items in a single request is 400.
+                /// So I think 300 should be fine by them.
+                let chunkedRecords = recordsToStore.chunkItUp(by: 300)
+                for chunk in chunkedRecords {
+                    self.syncRecordsToCloudKit(recordsToStore: chunk, recordIDsToDelete: recordIDsToDelete, completion: completion)
+                }
+            default:
+                return
+            }
+        }
+        
+        database.add(modifyOpe)
+    }
+    
+}
