+++ IceCream/Classes/PublicDatabaseManager.swift
@@ -1,0 +1,112 @@
+//
+//  PublicDatabaseManager.swift
+//  IceCream
+//
+//  Created by caiyue on 2019/4/22.
+//
+
+import CloudKit
+
+final class PublicDatabaseManager: DatabaseManager {
+    
+    let container: CKContainer
+    let database: CKDatabase
+    
+    let syncObjects: [Syncable]
+    
+    init(objects: [Syncable], container: CKContainer) {
+        self.syncObjects = objects
+        self.container = container
+        self.database = container.publicCloudDatabase
+    }
+    
+    func fetchChangesInDatabase(_ callback: (() -> Void)?) {
+        syncObjects.forEach { [weak self] syncObject in
+            let predicate = NSPredicate(value: true)
+            let query = CKQuery(recordType: syncObject.recordType, predicate: predicate)
+            let queryOperation = CKQueryOperation(query: query)
+            self?.excuteQueryOperation(queryOperation: queryOperation, on: syncObject, callback: callback)
+        }
+    }
+    
+    func createCustomZonesIfAllowed(_ completion: ((Error?) -> ())?) {
+        
+    }
+    
+    func createDatabaseSubscriptionIfHaveNot() {
+        syncObjects.forEach { createSubscriptionInPublicDatabase(on: $0) }
+    }
+    
+    func startObservingTermination() {
+        #if os(iOS) || os(tvOS)
+        
+        NotificationCenter.default.addObserver(self, selector: #selector(self.cleanUp), name: UIApplication.willTerminateNotification, object: nil)
+        
+        #elseif os(macOS)
+        
+        NotificationCenter.default.addObserver(self, selector: #selector(self.cleanUp), name: NSApplication.willTerminateNotification, object: nil)
+        
+        #endif
+    }
+    
+    func registerLocalDatabase() {
+        syncObjects.forEach {
+            $0.registerLocalDatabase()
+        }
+    }
+    
+    // MARK: - Private Methods
+    private func excuteQueryOperation(queryOperation: CKQueryOperation,on syncObject: Syncable, callback: (() -> Void)? = nil) {
+        queryOperation.recordFetchedBlock = { record in
+            syncObject.add(record: record)
+        }
+        
+        queryOperation.queryCompletionBlock = { [weak self] cursor, error in
+            guard let self = self else { return }
+            if let cursor = cursor {
+                let subsequentQueryOperation = CKQueryOperation(cursor: cursor)
+                self.excuteQueryOperation(queryOperation: subsequentQueryOperation, on: syncObject, callback: callback)
+                return
+            }
+            switch ErrorHandler.shared.resultType(with: error) {
+            case .success:
+                DispatchQueue.main.async {
+                    callback?()
+                }
+            case .retry(let timeToWait, _):
+                ErrorHandler.shared.retryOperationIfPossible(retryAfter: timeToWait, block: {
+                    self.excuteQueryOperation(queryOperation: queryOperation, on: syncObject, callback: callback)
+                })
+            default:
+                break
+            }
+        }
+        
+        database.add(queryOperation)
+    }
+    
+    private func createSubscriptionInPublicDatabase(on syncObject: Syncable) {
+        let predict = NSPredicate(value: true)
+        let subscription = CKQuerySubscription(recordType: syncObject.recordType, predicate: predict, subscriptionID: IceCreamConstant.cloudKitSubscriptionID, options: [.firesOnRecordCreation, .firesOnRecordUpdate, .firesOnRecordDeletion])
+        
+        let notificationInfo = CKSubscription.NotificationInfo()
+        notificationInfo.shouldSendContentAvailable = true // Silent Push
+        
+        subscription.notificationInfo = notificationInfo
+        
+        let createOp = CKModifySubscriptionsOperation(subscriptionsToSave: [subscription], subscriptionIDsToDelete: [])
+        createOp.modifySubscriptionsCompletionBlock = { _, _, error in
+            guard error == nil else { return }
+            // TODO: add subscription flag in public database manager
+//            self.subscriptionIsLocallyCached = true
+        }
+        createOp.qualityOfService = .utility
+        database.add(createOp)
+    }
+    
+    @objc func cleanUp() {
+        for syncObject in syncObjects {
+            syncObject.cleanUp()
+        }
+    }
+}
